tasklist

+++ app/Http/Controllers/TasksController.php
@@ -15,12 +15,27 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        
-        return view('tasks.index', [
-            'tasks' => $tasks
-            ]);
+   
+        $data = [];
+       if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+             // メッセージ一覧ビューでそれを表示
+             return view('tasks.index', [
+            'tasks' => $tasks,
+             ]);
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
     }
+    
 
     /**
      * Show the form for creating a new resource.
@@ -49,12 +64,15 @@ class TasksController extends Controller
             'content' => 'required|max:255',
             ]);
             
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+         // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+             ]);
+            
+       
         
-        return redirect('/');
+        return redirect('/'); 
     }
 
     /**
@@ -66,10 +84,12 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id) {
         
         return view('tasks.show', [
             'task' => $task,
             ]);
+        }
     }
 
     /**
@@ -82,9 +102,11 @@ class TasksController extends Controller
     {
         $task = Task::findOrFail($id);
         
+        if (\Auth::id() === $task->user_id) {
         return view('tasks.edit', [
             'task' => $task,
             ]);
+        }
     }
 
     /**
@@ -103,9 +125,13 @@ class TasksController extends Controller
             
         $task = Task::findOrFail($id);
         
+        if (\Auth::id() === $task->user_id) {
+            
+        
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
+        }
         
         return redirect('/');
     }
@@ -118,10 +144,12 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::findOrFail($id);
+        $task = \App\Task::findOrFail($id);
         
-        $task->delete();
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         
-        return redirect('/');
+        return redirect('/'); 
     }
 }
